wip update prediction/interpolation plugins
