+++ lightyear/src/client/config.rs
@@ -100,3 +100,9 @@ pub struct ClientConfig {
     pub interpolation: InterpolationConfig,
     pub replication: ReplicationConfig,
 }
+
+impl ClientConfig {
+    pub(crate) fn is_unified(&self) -> bool {
+        self.shared.unified
+    }
+}
