+++ lightyear/src/client/interpolation/mod.rs
@@ -10,6 +10,7 @@ pub use plugin::{add_interpolation_systems, add_prepare_interpolation_systems};
 pub use visual_interpolation::{VisualInterpolateStatus, VisualInterpolationPlugin};
 
 use crate::client::components::{Confirmed, LerpFn, SyncComponent};
+use crate::client::config::ClientConfig;
 use crate::client::connection::ConnectionManager;
 use crate::client::interpolation::resource::InterpolationManager;
 use crate::protocol::Protocol;
@@ -20,6 +21,7 @@ mod interpolate;
 pub mod interpolation_history;
 pub mod plugin;
 mod resource;
+mod spawn;
 mod visual_interpolation;
 
 /// Interpolator that performs linear interpolation.
@@ -52,53 +54,4 @@ pub struct Interpolated {
     //  - despawn immediately all components
     //  - leave the entity alive until the confirmed entity catches up to it and then it gets removed.
     //    - or do this only for certain components (audio, animation, particles..) -> mode on PredictedComponent
-}
-
-pub fn spawn_interpolated_entity<P: Protocol>(
-    connection: Res<ConnectionManager<P>>,
-    mut manager: ResMut<InterpolationManager>,
-    mut commands: Commands,
-    mut confirmed_entities: Query<(Entity, Option<&mut Confirmed>), Added<ShouldBeInterpolated>>,
-) {
-    for (confirmed_entity, confirmed) in confirmed_entities.iter_mut() {
-        // spawn a new interpolated entity
-        let interpolated_entity_mut = commands.spawn(Interpolated { confirmed_entity });
-        let interpolated = interpolated_entity_mut.id();
-
-        // update the entity mapping
-        manager
-            .interpolated_entity_map
-            .confirmed_to_interpolated
-            .insert(confirmed_entity, interpolated);
-
-        // add Confirmed to the confirmed entity
-        // safety: we know the entity exists
-        let mut confirmed_entity_mut = commands.get_entity(confirmed_entity).unwrap();
-        if let Some(mut confirmed) = confirmed {
-            confirmed.interpolated = Some(interpolated);
-        } else {
-            // get the confirmed tick for the entity
-            // if we don't have it, something has gone very wrong
-            let confirmed_tick = connection
-                .replication_receiver
-                .get_confirmed_tick(confirmed_entity)
-                .unwrap();
-            confirmed_entity_mut.insert(Confirmed {
-                interpolated: Some(interpolated),
-                predicted: None,
-                tick: confirmed_tick,
-            });
-        }
-        trace!(
-            "Spawn interpolated entity {:?} for confirmed: {:?}",
-            interpolated,
-            confirmed_entity
-        );
-        #[cfg(feature = "metrics")]
-        {
-            metrics::counter!("spawn_interpolated_entity")
-                .increment(1)
-                .increment(1);
-        }
-    }
 }
